requirements.txt
pinecone>=5.0.0
openai>=1.0.0
python-dotenv==1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
python-docx>=1.1.0
pypdf2>=3.0.0

